revise linechart

+++ app/src/main/java/com/example/healthhelper/person/model/WeightData.kt
@@ -3,7 +3,7 @@ package com.example.healthhelper.person.model
 
 data class WeightData(
     val date: String,
-    val weight: Double,
-    val bmi: Double,
-    val bodyFat: Double,
+    val weight: Float,
+    val bmi: Float,
+    val bodyFat: Float,
 )
